Remove `File::getContents()`; use `FilesystemReader` in `FileSymbolScanner`

+++ tests/Integration/Util/IntegrationTestCase.php
@@ -132,6 +132,10 @@ class IntegrationTestCase extends TestCase
             }
         }
 
+        if (!is_dir($dir)) {
+            return;
+        }
+
         $filesystem->deleteDirectory($dir);
     }
 }
